refactor: rename QueryBuilder to SqlQueryBuilder
