+++ packages/backend/src/utils/QueryBuilder/SqlQueryBuilder.test.ts
@@ -3,7 +3,7 @@ import {
     DimensionType,
     FilterOperator,
 } from '@lightdash/common';
-import { QueryBuilder } from './queryBuilder';
+import { SqlQueryBuilder } from './SqlQueryBuilder';
 import {
     DEFAULT_CONFIG,
     MULTI_FIELD_REFERENCE_MAP,
@@ -17,12 +17,12 @@ import {
     SIMPLE_FILTER_RULE,
     SIMPLE_QUERY_SQL,
     SIMPLE_REFERENCE_MAP,
-} from './queryBuilder.class.mocks';
+} from './SqlQueryBuilder.mocks';
 
-describe('QueryBuilder class', () => {
+describe('SqlQueryBuilder class', () => {
     describe('constructor', () => {
         it('should initialize with minimal arguments', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: {},
                     select: [],
@@ -36,7 +36,7 @@ describe('QueryBuilder class', () => {
         });
 
         it('should initialize with all arguments', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: SIMPLE_REFERENCE_MAP,
                     select: ['test_field'],
@@ -56,7 +56,7 @@ describe('QueryBuilder class', () => {
 
     describe('SQL generation', () => {
         it('should generate correct SQL for a simple query', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: SIMPLE_REFERENCE_MAP,
                     select: ['test_field'],
@@ -69,7 +69,7 @@ describe('QueryBuilder class', () => {
         });
 
         it('should generate correct SQL for a query with filters', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: SIMPLE_REFERENCE_MAP,
                     select: ['test_field'],
@@ -86,7 +86,7 @@ describe('QueryBuilder class', () => {
         });
 
         it('should generate correct SQL for a query with two filters', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: SIMPLE_REFERENCE_MAP,
                     select: ['test_field'],
@@ -103,7 +103,7 @@ describe('QueryBuilder class', () => {
         });
 
         it('should generate correct SQL for a query with a subquery in FROM', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: {
                         test_field: {
@@ -124,7 +124,7 @@ describe('QueryBuilder class', () => {
         });
 
         it('should handle SQL with semicolons and comments in FROM', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: {
                         test_field: {
@@ -151,7 +151,7 @@ describe('QueryBuilder class', () => {
 
     describe('complex scenarios', () => {
         it('should generate SQL with nested filter groups', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: MULTI_FIELD_REFERENCE_MAP,
                     select: ['field1', 'field2', 'field3'],
@@ -198,7 +198,7 @@ describe('QueryBuilder class', () => {
         });
 
         it('should handle empty select list by selecting all columns', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: {},
                     select: [],
@@ -213,7 +213,7 @@ describe('QueryBuilder class', () => {
 
     describe('error handling', () => {
         it('should throw an error for unknown reference', () => {
-            const queryBuilder = new QueryBuilder(
+            const queryBuilder = new SqlQueryBuilder(
                 {
                     referenceMap: {},
                     select: ['unknown_field'],
